+++ phpBB2/admin/admin_user_register.php
@@ -222,8 +222,9 @@ if ( isset($_POST['submit']) )
 
 		$new_password = phpbb_password_hash($new_password);
 
-		$sql = "INSERT INTO " . USERS_TABLE . "	(user_id, username, user_regdate, user_password, user_email, user_style, user_timezone, user_dateformat, user_lang, user_level, user_active, user_actkey)
-			VALUES ($user_id, '" . $db->sql_escape($username) . "',	" . time() . ",	'" . $db->sql_escape($new_password) . "',	'" . $db->sql_escape($email) . "', $user_style, $user_timezone, '" . $db->sql_escape($user_dateformat) . "', '" . $db->sql_escape($user_lang) . "', 0, 1, 'user_actkey')";
+		$account_created_at = time();
+		$sql = "INSERT INTO " . USERS_TABLE . "	(user_id, username, user_regdate, user_password, user_email, user_style, user_timezone, user_dateformat, user_lang, user_level, user_active, user_actkey, user_passwd_change, ct_last_pw_change)
+			VALUES ($user_id, '" . $db->sql_escape($username) . "',	" . $account_created_at . ",	'" . $db->sql_escape($new_password) . "',	'" . $db->sql_escape($email) . "', $user_style, $user_timezone, '" . $db->sql_escape($user_dateformat) . "', '" . $db->sql_escape($user_lang) . "', 0, 1, 'user_actkey', " . $account_created_at . ", " . $account_created_at . ")";
 		if ( !($result = $db->sql_query($sql, BEGIN_TRANSACTION)) )
 		{
 			message_die(GENERAL_ERROR, 'Could not insert data into users table', '', __LINE__, __FILE__, $sql);